correction on the finance_calculator.pycode

+++ python_introduction/finance_calculator.py
@@ -1,11 +1,10 @@
-monthly_income = int(input ("Enter your monthly income: "))
+monthly_income = (input ("Enter your monthly income: "))
 
-monthly_expence = int(input ("Enter your total monthly expenses: "))
+monthly_expence = float(input ("Enter your total monthly expenses: "))
 
 monthly_savings = monthly_income - monthly_expence
 
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
-
 
 print (f"Your monthly savings are ${monthly_savings}")
 print(f"Projected savings after one year, with interest, is: ${projected_savings}.")
